+++ src/utils/dateUtils/get-days-from-month.ts
@@ -1,0 +1,2 @@
+export const getDaysFromMonth = (year: number, month: number) =>
+  new Date(year, month, 0).getDate();
