+++ src/utils/dateUtils/is-dates-equal.ts
@@ -1,0 +1,2 @@
+export const isDatesEqual = (date1: Date, date2: Date) =>
+  date1?.toString() === date2?.toString();
