+++ src/utils/dateUtils/is-day-belongs-to-month.ts
@@ -1,0 +1,2 @@
+export const isDayBelongsToMonth = (date: Date, month: number) =>
+  date.getMonth() === month - 1;
