+++ src/utils/dateUtils/next-date.ts
@@ -1,0 +1,5 @@
+export const nextDate = (date: Date, value: number) => {
+  const next = new Date(date);
+  next.setDate(date.getDate() + value);
+  return next;
+};
