+++ src/utils/inputUtils/parse-input-date.ts
@@ -1,0 +1,7 @@
+import { splitDate } from "./split-date";
+import { zeroPrefix } from "./zero-prefix";
+
+export const parseInputDate = (value: Date) => {
+  const { day, month, year } = splitDate(value);
+  return `${zeroPrefix(day)}/${zeroPrefix(month)}/${year}`;
+};
