+++ src/utils/inputUtils/split-date.ts
@@ -1,0 +1,8 @@
+export const splitDate = (date: Date) => {
+  const [day, month, year] = [
+    date.getDate(),
+    date.getMonth() + 1,
+    date.getFullYear(),
+  ];
+  return { day, month, year };
+};
