+++ src/utils/inputUtils/split-string.ts
@@ -1,0 +1,4 @@
+export const splitString = (value: string) => {
+  const [day, month, year] = value.split("/").map((item) => parseInt(item));
+  return { day, month, year };
+};
