+++ src/utils/inputUtils/zero-prefix.ts
@@ -1,0 +1,1 @@
+export const zeroPrefix = (num: number) => (num < 10 ? `0${num}` : num);
